+++ clutter/clutter/clutter-blur.c
@@ -242,8 +242,8 @@ create_fbo (ClutterBlur *blur,
   scaled_height = floorf (height / blur->downscale_factor);
 
   pass->texture = cogl_texture_2d_new_with_size (ctx,
-                                                 scaled_width,
-                                                 scaled_height);
+                                                 (int) scaled_width,
+                                                 (int) scaled_height);
   if (!pass->texture)
     return FALSE;
 
@@ -354,19 +354,19 @@ clutter_blur_new (CoglTexture *texture,
   BlurPass *vpass;
 
   g_return_val_if_fail (texture != NULL, NULL);
-  g_return_val_if_fail (radius >= 0.0, NULL);
+  g_return_val_if_fail (radius >= 0.0f, NULL);
 
   width = cogl_texture_get_width (texture);
   height = cogl_texture_get_height (texture);
 
   blur = g_new0 (ClutterBlur, 1);
-  blur->sigma = radius / 2.0;
+  blur->sigma = radius / 2.0f;
   blur->source_texture = g_object_ref (texture);
   blur->downscale_factor = calculate_downscale_factor (width,
                                                        height,
                                                        blur->sigma);
 
-  if (G_APPROX_VALUE (blur->sigma, 0.0, FLT_EPSILON))
+  if (G_APPROX_VALUE (blur->sigma, 0.0f, FLT_EPSILON))
     goto out;
 
   vpass = &blur->pass[VERTICAL];
